# Small changes for performance