I'm trying to fix #5. I don't think this is the best way. I'd like to discuss.

I think what is happening is that it takes more than 5 seconds to run, so the calls pile up if it is scheduled every 5 seconds. Ideally it would trigger a new call only after the current call finishes, rather than running on any schedule - 5 seconds or 20 seconds or otherwise.

The `.endwith` change is basically grasping at straws.